Share database-backed Scout engine behavior
